// src/types/nansen/api.gen.ts
export const S_PaginationInfo = z.looseObject({
  "page": z.number().nullish(),
  "per_page": z.number().nullish(),
  "is_last_page": z.boolean().nullish(),
});
// ...
export const S_TGMDexTrade = z.looseObject({
  "block_timestamp": z.string(),
  "transaction_hash": z.string(),
  "trader_address": z.string(),
  "trader_address_label": z.string().nullish(),
  "action": S_TGMWhoBoughtSoldType,
  "token_address": z.string(),
  "token_name": z.string(),
  "token_amount": z.number(),
  "traded_token_address": z.string(),
  "traded_token_name": z.string(),
  "traded_token_amount": z.number(),
  "estimated_swap_price_usd": z.number(),
  "estimated_value_usd": z.number(),
});
// ...
export const S_TGMDexTradesResponse = z.looseObject({
  "data": z.array(S_TGMDexTrade),
  "pagination": S_PaginationInfo,
});
export type TGMDexTradesResponse = z.infer<typeof S_TGMDexTradesResponse>;
// ...
export const S_TGMPnlLeaderboard = z.looseObject({
  "trader_address": z.string(),
  "trader_address_label": z.string().nullish(),
  "price_usd": z.number().nullish(),
  "pnl_usd_realised": z.number().nullish(),
  "pnl_usd_unrealised": z.number().nullish(),
  "holding_amount": z.number().nullish(),
  "holding_usd": z.number().nullish(),
  "max_balance_held": z.number().nullish(),
  "max_balance_held_usd": z.number().nullish(),
  "still_holding_balance_ratio": z.number().nullish(),
  "netflow_amount_usd": z.number().nullish(),
  "netflow_amount": z.number().nullish(),
  "roi_percent_total": z.number().nullish(),
  "roi_percent_realised": z.number().nullish(),
  "roi_percent_unrealised": z.number().nullish(),
  "pnl_usd_total": z.number().nullish(),
  "nof_trades": z.number().nullish(),
});
// ...
export const S_TGMJupDca = z.looseObject({
  "since_timestamp": z.string().nullish(),
  "last_timestamp": z.string().nullish(),
  "trader_address": z.string(),
  "creation_hash": z.string().nullish(),
  "trader_label": z.string().nullish(),
  "dca_vault_address": z.string().nullish(),
  "input_mint_address": z.string().nullish(),
  "output_mint_address": z.string().nullish(),
  "deposit_amount": z.number().nullish(),
  "deposit_spent": z.number().nullish(),
  "other_token_redeemed": z.number().nullish(),
  "status": S_TGMJupDcaStatus.nullish(),
  "token_input": z.string().nullish(),
  "token_output": z.string().nullish(),
  "deposit_usd_value": z.number().nullish(),
});
// ...
export const S_TGMJupDcaResponse = z.looseObject({
  "data": z.array(S_TGMJupDca),
  "pagination": S_PaginationInfo,
});
export type TGMJupDcaResponse = z.infer<typeof S_TGMJupDcaResponse>;
// ...
export const S_TGMPerpPosition = z.looseObject({
  "address": z.string().nullish(),
  "address_label": z.string().nullish(),
  "side": S_PositionSide.nullish(),
  "position_value_usd": z.number().nullish(),
  "position_size": z.number().nullish(),
  "leverage": z.string(),
  "leverage_type": z.string().nullish(),
  "entry_price": z.number().nullish(),
  "mark_price": z.number().nullish(),
  "liquidation_price": z.number().nullish(),
  "funding_usd": z.number().nullish(),
  "upnl_usd": z.number().nullish(),
});
// ...
export const S_TGMPerpPositionsResponse = z.looseObject({
  "data": z.array(S_TGMPerpPosition),
  "pagination": S_PaginationInfo,
});
export type TGMPerpPositionsResponse = z.infer<typeof S_TGMPerpPositionsResponse>;
// ...
export const S_TGMPnlLeaderboardResponse = z.looseObject({
  "data": z.array(S_TGMPnlLeaderboard),
  "pagination": S_PaginationInfo,
});
export type TGMPnlLeaderboardResponse = z.infer<typeof S_TGMPnlLeaderboardResponse>;
export const S_TGMPositionIntelligence = z.looseObject({
  "smart_trader_longs_usd": z.number().nullish(),
  "smart_trader_shorts_usd": z.number().nullish(),
  "smart_trader_total_usd": z.number().nullish(),
  "whale_longs_usd": z.number().nullish(),
  "whale_shorts_usd": z.number().nullish(),
  "whale_total_usd": z.number().nullish(),
  "public_figure_longs_usd": z.number().nullish(),
  "public_figure_shorts_usd": z.number().nullish(),
  "public_figure_total_usd": z.number().nullish(),
});
export type TGMPositionIntelligence = z.infer<typeof S_TGMPositionIntelligence>;
export const S_TGMPositionIntelligenceRequest = z.looseObject({
  "token_address": z.string(),
});
export type TGMPositionIntelligenceRequest = z.infer<typeof S_TGMPositionIntelligenceRequest>;
export const S_TGMPositionIntelligenceResponse = z.looseObject({
  "data": z.array(S_TGMPositionIntelligence),
});
export type TGMPositionIntelligenceResponse = z.infer<typeof S_TGMPositionIntelligenceResponse>;
// ...
export const S_TGMTransfer = z.looseObject({
  "block_timestamp": z.string(),
  "transaction_hash": z.string(),
  "from_address": z.string(),
  "to_address": z.string(),
  "from_address_label": z.string().nullish(),
  "to_address_label": z.string().nullish(),
  "transaction_type": z.string().nullish(),
  "transfer_amount": z.number().nullish(),
  "transfer_value_usd": z.number().nullish(),
});
// ...
export const S_TGMTransfersResponse = z.looseObject({
  "data": z.array(S_TGMTransfer),
  "pagination": S_PaginationInfo,
});
export type TGMTransfersResponse = z.infer<typeof S_TGMTransfersResponse>;